+++ contracts/interfaces/IParentRollingRewarder.sol
@@ -4,5 +4,5 @@ pragma solidity 0.8.23;
 import "./IRewarder.sol";
 
 interface IParentRollingRewarder is IRewarder {
-    function initialize(uint256 _poolId) external;
+    function initialize(uint8 _poolId) external;
 }
